PRM-1577 Refactored
