+++ app/controllers/PenaltySelectionController.scala
@@ -28,11 +28,13 @@ import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendController
 import utils.SessionKeys
 import views.html.{AppealSinglePenaltyPage, PenaltySelectionPage}
+import views.html.{AppealCoverBothPenaltiesPage, PenaltySelectionPage}
 import viewtils.RadioOptionHelper
 
 import javax.inject.Inject
 
 class PenaltySelectionController @Inject()(penaltySelectionPage: PenaltySelectionPage,
+                                           appealCoverBothPenaltiesPage: AppealCoverBothPenaltiesPage,
                                            appealSinglePenaltyPage: AppealSinglePenaltyPage,
                                            navigation: Navigation)
                                           (implicit mcc: MessagesControllerComponents,
@@ -74,6 +76,12 @@ class PenaltySelectionController @Inject()(penaltySelectionPage: PenaltySelectio
       val nextPageUrl: String = navigation.nextPage(AppealSinglePenaltyPage, mode).url
       //TODO: penaltyInformation should be populated
       Ok(appealSinglePenaltyPage(pageMode(AppealSinglePenaltyPage, mode), nextPageUrl, "’penalty information here’"))
+    }
+  }
+
+  def onPageLoadForAppealCoverBothPenalties(): Action[AnyContent] = (authorise andThen dataRequired) {
+    implicit request => {
+      Ok(appealCoverBothPenaltiesPage())
     }
   }
 }
