--- it/controllers/AppealCoverPenaltiesControllerISpec.scala
@@ -1,45 +1,0 @@
-/*
- * Copyright 2022 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package controllers
-
-import play.api.mvc.AnyContent
-import play.api.test.FakeRequest
-import play.api.test.Helpers._
-import uk.gov.hmrc.http.SessionKeys.authToken
-import utils.{IntegrationSpecCommonBase, SessionKeys}
-
-class AppealCoverPenaltiesControllerISpec extends IntegrationSpecCommonBase{
-  val controller: AppealCoverPenaltiesController = injector.instanceOf[AppealCoverPenaltiesController]
-
-  "GET /appeal-cover-for-both-penalties" should {
-    "return 200 (OK) when the user is authorised" in {
-      val fakeRequestWithCorrectKeys: FakeRequest[AnyContent] = FakeRequest("GET", "/appeal-cover-for-both-penalties").withSession(
-        authToken -> "1234",
-        (SessionKeys.penaltyNumber, "1234"),
-        (SessionKeys.appealType, "Late_Submission"),
-        (SessionKeys.startDateOfPeriod, "2020-01-01"),
-        (SessionKeys.endDateOfPeriod, "2020-01-01"),
-        (SessionKeys.dueDateOfPeriod, "2020-02-07"),
-        (SessionKeys.dateCommunicationSent, "2020-02-08"),
-        (SessionKeys.journeyId, "1234")
-      )
-
-      val request = await(controller.onPageLoadForAppealCoverBothPenalties()(fakeRequestWithCorrectKeys))
-      request.header.status shouldBe OK
-    }
-  }
-}
